# openapi.yaml
openapi: 3.0.0
info:
  title: Flyway API
  description: API for Flyway application
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local development server
...
components:
  schemas:
...
    CreateOrganizationRequest:
      type: object
      description: Request to create a new organization
      required:
        - name
      properties:
        name:
          type: string
          description: Organization name
        description:
          type: string
          description: Organization description
    UpdateOrganizationRequest:
      type: object
      description: Request to update an organization
      properties:
        name:
          type: string
          description: Organization name
        description:
          type: string
          description: Organization description
...
    CreateMobileApplicationRequest:
      type: object
      description: Request to create a new mobile application
      required:
        - bundleId
        - name
      properties:
        bundleId:
          type: string
          description: Unique bundle identifier for the mobile application (e.g., com.example.app)
        name:
          type: string
          description: Mobile application name
        description:
          type: string
          description: Mobile application description
    UpdateMobileApplicationRequest:
      type: object
      description: Request to update a mobile application
      required:
        - name
      properties:
        name:
          type: string
          description: Mobile application name
        description:
          type: string
          description: Mobile application description
...
    OrganizationResponse:
      type: object
      description: Organization information
      required:
        - id
        - name
        - createdAt
      properties:
        id:
          type: string
          description: Organization ID
        name:
          type: string
          description: Organization name
        description:
          type: string
          description: Organization description
        createdAt:
          type: string
          format: date-time
          description: When the organization was created
        updatedAt:
          type: string
          format: date-time
          description: When the organization was last updated
...
    MobileApplicationResponse:
      type: object
      description: Mobile application information
      required:
        - id
        - bundleId
        - organizationId
        - name
        - createdBy
        - createdAt
      properties:
        id:
          type: string
          description: Mobile application ID
        bundleId:
          type: string
          description: Bundle identifier for the mobile application
        organizationId:
          type: string
          description: Organization ID that owns this application
        name:
          type: string
          description: Mobile application name
        description:
          type: string
          description: Mobile application description
        createdBy:
          type: string
          description: User ID who created the application
        createdAt:
          type: string
          format: date-time
          description: When the application was created
        updatedAt:
          type: string
          format: date-time
          description: When the application was last updated
...
    ErrorResponse:
      type: object
      description: Error response
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error message
        details:
          type: string
          description: Error details
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
  responses:
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Unauthorized access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Conflict - resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /organizations:
    get:
      operationId: getCurrentOrganization
      summary: Get current organization
      tags:
        - Organizations
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Organization retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    
    post:
      operationId: createOrganization
      summary: Create a new organization
      tags:
        - Organizations
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationRequest'
      responses:
        '201':
          description: Organization created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    
    put:
      operationId: updateCurrentOrganization
      summary: Update current organization
      tags:
        - Organizations
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrganizationRequest'
      responses:
        '200':
          description: Organization updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizationResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /mobile-applications:
    get:
      operationId: getAllMobileApplications
      summary: Get all mobile applications for the organization
      tags:
        - Mobile Applications
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Mobile applications retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MobileApplicationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
    
    post:
      operationId: createMobileApplication
      summary: Create a new mobile application
      tags:
        - Mobile Applications
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMobileApplicationRequest'
      responses:
        '201':
          description: Mobile application created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileApplicationResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /mobile-applications/bundle/{bundleId}:
    get:
      operationId: getMobileApplicationByBundleId
      summary: Get mobile application by bundle ID
      tags:
        - Mobile Applications
      security:
        - BearerAuth: []
      parameters:
        - name: bundleId
          in: path
          required: true
          schema:
            type: string
          description: The bundle ID of the mobile application (e.g., com.example.app)
      responses:
        '200':
          description: Mobile application retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileApplicationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /mobile-applications/{id}:
    get:
      operationId: getMobileApplicationById
      summary: Get mobile application by ID
      tags:
        - Mobile Applications
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Mobile application retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileApplicationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    
    put:
      operationId: updateMobileApplication
      summary: Update mobile application
      tags:
        - Mobile Applications
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMobileApplicationRequest'
      responses:
        '200':
          description: Mobile application updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileApplicationResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    
    delete:
      operationId: deleteMobileApplication
      summary: Delete mobile application
      tags:
        - Mobile Applications
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Mobile application deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...